{{cookiecutter.project_name}}/requirements.txt
# {{cookiecutter.project_name}} Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Service discovery
python-consul==1.1.0

# Database (optional - uncomment as needed)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
